+++ src/main/java/net/geforcemods/scguardgolem/SCGContent.java
@@ -1,6 +1,7 @@
 package net.geforcemods.scguardgolem;
 
 import net.geforcemods.scguardgolem.entity.SecurityGolemEntity;
+import net.geforcemods.scguardgolem.inventory.GolemMenu;
 import net.geforcemods.scguardgolem.item.SCGManualItem;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.Registries;
@@ -11,12 +12,14 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.animal.golem.IronGolem;
+import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.common.extensions.IMenuTypeExtension;
 import net.neoforged.neoforge.event.entity.EntityAttributeCreationEvent;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredItem;
@@ -31,6 +34,8 @@ public class SCGContent {
             DeferredRegister.createItems(SCGuardGolem.MODID);
     public static final DeferredRegister<CreativeModeTab> CREATIVE_MODE_TABS =
             DeferredRegister.create(Registries.CREATIVE_MODE_TAB, SCGuardGolem.MODID);
+    public static final DeferredRegister<MenuType<?>> MENU_TYPES =
+            DeferredRegister.create(Registries.MENU, SCGuardGolem.MODID);
 
     public static final DeferredHolder<EntityType<?>, EntityType<SecurityGolemEntity>> SECURITY_GOLEM =
             ENTITY_TYPES.register("security_golem", () ->
@@ -41,6 +46,10 @@ public class SCGContent {
                             .setShouldReceiveVelocityUpdates(true)
                             .build(ResourceKey.create(Registries.ENTITY_TYPE,
                                     Identifier.fromNamespaceAndPath(SCGuardGolem.MODID, "security_golem"))));
+
+    public static final DeferredHolder<MenuType<?>, MenuType<GolemMenu>> GOLEM_MENU =
+            MENU_TYPES.register("golem_menu", () ->
+                    IMenuTypeExtension.create(GolemMenu::new));
 
     public static final DeferredItem<SCGManualItem> SCG_MANUAL =
             ITEMS.registerItem("scg_manual", SCGManualItem::new,
@@ -60,6 +69,7 @@ public class SCGContent {
         ENTITY_TYPES.register(modBus);
         ITEMS.register(modBus);
         CREATIVE_MODE_TABS.register(modBus);
+        MENU_TYPES.register(modBus);
     }
 
     @SubscribeEvent